@@ -1,6 +1,6 @@
-﻿using BenchmarkDotNet.Running;
+﻿namespace Kasp.ObjectMapper.Benchmark {
+	using BenchmarkDotNet.Running;
 
-namespace Kasp.ObjectMapper.Benchmark {
 	class Program {
 		static void Main(string[] args) {
 			BenchmarkRunner.Run<MapObjectBenchmark>();
